+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [2.16.1-lts.3] - 2022-12-XX
+-   Update `FileViewer` Class to avoid the property ContactsService to be serialized when the Object is used with JSON.stringify tool.
+
 ## [2.16.1-lts.2] - 2022-11-18
 -   Fix events treatment `ConversationHistoryHandler::onHistoryMessageReceived` when there are several times the tag `headers`. These events are raised as result of `conversations::getHistoryPage` API.
 -   Add a new header with the rainbow application id `x-rainbow-client-id` in requests.